+++ StackOverflowLab/Controllers/QuestionsController.cs
@@ -44,17 +44,35 @@ namespace StackOverflowLab.Controllers
 
         public IActionResult DeleteQuestion(int questid)
         {
-            DAL.DeleteQuestion(questid);
-            return Redirect("/Questions");
+            Question question = DAL.GetQuestion(questid);
+            if (question.username == DAL.CurrentUser || DAL.CurrentUser == "Admin")
+            {
+                DAL.DeleteQuestion(questid);
+                return Redirect("/Questions");
+            }
+            else
+            {
+                TempData["DeleteQuestionUsername"] = "*You cannot delete somebody elses question.";
+                return Redirect("/Questions");
+            }
         }
         public IActionResult EditQuestionForm(int questid)
         {
-            Question quest = DAL.GetQuestion(questid);
-            return View(quest);
- 
+            Question question = DAL.GetQuestion(questid);
+            if (question.username == DAL.CurrentUser || DAL.CurrentUser == "Admin")
+            {
+                return View(question);
+            }
+            else
+            {
+                TempData["EditQuestionUsername"] = "*You cannot edit somebody elses question.";
+                return Redirect("/Questions");
+            }
         }
         public IActionResult SaveQuestion(Question question)
         {
+            question.username = DAL.CurrentUser;
+            question.posted = DateTime.Now;
             DAL.UpdateQuestion(question);
             return Redirect("/Questions/index");
         }
